Move content api route into its own file
